+++ app/pages/sections/catalogue-solutions/catalogue-solutions/ui-tests/general.ui.test.js
@@ -131,7 +131,7 @@ test('should render the Add Catalogue Solution button', async (t) => {
     .expect(await extractInnerText(addSolutionButton)).eql(content.addSolutionButtonText);
 });
 
-test('should navigate to /organisation/order-id/catalogue-solutions/select when Add Catalogue Solution button is clicked', async (t) => {
+test('should navigate to /organisation/order-id/catalogue-solutions/select/solution when Add Catalogue Solution button is clicked', async (t) => {
   await pageSetup();
   await t.navigateTo(pageUrl);
 
@@ -139,7 +139,7 @@ test('should navigate to /organisation/order-id/catalogue-solutions/select when 
 
   await t
     .click(addSolutionButton)
-    .expect(getLocation()).eql('http://localhost:1234/order/organisation/order-id/catalogue-solutions/solution');
+    .expect(getLocation()).eql('http://localhost:1234/order/organisation/order-id/catalogue-solutions/select/solution');
 });
 
 test('should render the Continue button', async (t) => {
