+++ app/pages/sections/catalogue-solutions/routes.js
@@ -6,26 +6,8 @@ import {
   getCatalogueSolutionsPageContext,
   putCatalogueSolutions,
 } from './catalogue-solutions/controller';
-import {
-  findSolutions,
-  getSupplierId,
-  getSolutionsErrorPageContext,
-  getSolutionsPageContext,
-  validateSolutionForm,
-} from './select/solution/controller';
-import {
-  findSolutionPrices,
-  getSolutionPriceErrorPageContext,
-  getSolutionPricePageContext,
-  validateSolutionPriceForm,
-} from './select/price/controller';
-import {
-  getRecipientPageContext,
-  getRecipients,
-  getSolution,
-  validateRecipientForm,
-  getRecipientErrorPageContext,
-} from './select/recipient/controller';
+import { catalogueSolutionsSelectRoutes } from './select/routes';
+
 
 const router = express.Router({ mergeParams: true });
 
@@ -53,118 +35,7 @@ export const catalogueSolutionsRoutes = (authProvider, addContext, sessionManage
     return res.redirect(`${config.baseUrl}/organisation/${orderId}`);
   }));
 
-  router.get('/solution', authProvider.authorise({ claim: 'ordering' }), withCatch(authProvider, async (req, res) => {
-    const { orderId } = req.params;
-    const accessToken = extractAccessToken({ req, tokenType: 'access' });
-
-    const supplierId = await getSupplierId({ orderId, accessToken });
-    const solutions = await findSolutions({ supplierId, accessToken });
-    sessionManager.saveToSession({ req, key: 'solutions', value: solutions });
-
-    const context = await getSolutionsPageContext({ orderId, solutions });
-
-    logger.info(`navigating to order ${orderId} catalogue-solutions select solution page`);
-    return res.render('pages/sections/catalogue-solutions/select/solution/template.njk', addContext({ context, user: req.user, csrfToken: req.csrfToken() }));
-  }));
-
-  router.post('/solution', authProvider.authorise({ claim: 'ordering' }), withCatch(authProvider, async (req, res) => {
-    const { orderId } = req.params;
-
-    const response = validateSolutionForm({ data: req.body });
-
-    if (response.success) {
-      sessionManager.saveToSession({ req, key: 'selectedSolutionId', value: req.body.selectSolution });
-      logger.info('redirecting catalogue solutions select price page');
-      return res.redirect(`${config.baseUrl}/organisation/${orderId}/catalogue-solutions/solution/price`);
-    }
-
-    const solutions = sessionManager.getFromSession({ req, key: 'solutions' });
-    const context = await getSolutionsErrorPageContext({
-      orderId,
-      solutions,
-      validationErrors: response.errors,
-    });
-
-    return res.render('pages/sections/catalogue-solutions/select/solution/template.njk', addContext({ context, user: req.user, csrfToken: req.csrfToken() }));
-  }));
-
-  router.get('/solution/price', authProvider.authorise({ claim: 'ordering' }), withCatch(authProvider, async (req, res) => {
-    const { orderId } = req.params;
-    const accessToken = extractAccessToken({ req, tokenType: 'access' });
-
-    const solutionId = sessionManager.getFromSession({ req, key: 'selectedSolutionId' });
-    const solutionPrices = await findSolutionPrices({ solutionId, accessToken });
-    sessionManager.saveToSession({ req, key: 'solutionPrices', value: solutionPrices });
-
-    const context = await getSolutionPricePageContext({ orderId, solutionPrices });
-
-    logger.info(`navigating to order ${orderId} catalogue-solutions select price page`);
-    return res.render('pages/sections/catalogue-solutions/select/price/template.njk', addContext({ context, user: req.user, csrfToken: req.csrfToken() }));
-  }));
-
-  router.post('/solution/price', authProvider.authorise({ claim: 'ordering' }), withCatch(authProvider, async (req, res) => {
-    const { orderId } = req.params;
-
-    const response = validateSolutionPriceForm({ data: req.body });
-    if (response.success) {
-      sessionManager.saveToSession({ req, key: 'selectedPriceId', value: req.body.selectSolutionPrice });
-      logger.info('redirecting catalogue solutions select recipient page');
-      return res.redirect(`${config.baseUrl}/organisation/${orderId}/catalogue-solutions/solution/price/recipient`);
-    }
-
-    const solutionPrices = sessionManager.getFromSession({ req, key: 'solutionPrices' });
-    const context = await getSolutionPriceErrorPageContext({
-      orderId,
-      solutionPrices,
-      validationErrors: response.errors,
-    });
-
-    return res.render('pages/sections/catalogue-solutions/select/price/template.njk', addContext({ context, user: req.user, csrfToken: req.csrfToken() }));
-  }));
-
-  router.get('/solution/price/recipient', authProvider.authorise({ claim: 'ordering' }), withCatch(authProvider, async (req, res) => {
-    const { orderId } = req.params;
-    const accessToken = extractAccessToken({ req, tokenType: 'access' });
-
-    const solutionId = sessionManager.getFromSession({ req, key: 'selectedSolutionId' });
-    const solutionData = await getSolution({ solutionId });
-
-    const recipients = await getRecipients({ orderId, accessToken });
-    sessionManager.saveToSession({ req, key: 'recipients', value: recipients });
-
-    const context = await getRecipientPageContext({
-      orderId,
-      solutionName: solutionData.name,
-      recipients,
-    });
-
-    logger.info(`navigating to order ${orderId} catalogue-solutions select recipient page`);
-    return res.render('pages/sections/catalogue-solutions/select/recipient/template.njk', addContext({ context, user: req.user, csrfToken: req.csrfToken() }));
-  }));
-
-  router.post('/solution/price/recipient', authProvider.authorise({ claim: 'ordering' }), withCatch(authProvider, async (req, res) => {
-    const { orderId } = req.params;
-
-    const response = validateRecipientForm({ data: req.body });
-    if (response.success) {
-      sessionManager.saveToSession({ req, key: 'selectedRecipientId', value: req.body.selectRecipient });
-      logger.info('Redirect to new solution page');
-      return res.redirect(`${config.baseUrl}/organisation/${orderId}/catalogue-solutions/newsolution`);
-    }
-
-    const solutionId = sessionManager.getFromSession({ req, key: 'selectedSolutionId' });
-    const solutionData = await getSolution({ solutionId });
-
-    const recipients = sessionManager.getFromSession({ req, key: 'recipients' });
-    const context = await getRecipientErrorPageContext({
-      orderId,
-      solutionName: solutionData.name,
-      recipients,
-      validationErrors: response.errors,
-    });
-
-    return res.render('pages/sections/catalogue-solutions/select/recipient/template.njk', addContext({ context, user: req.user, csrfToken: req.csrfToken() }));
-  }));
+  router.use('/select', catalogueSolutionsSelectRoutes(authProvider, addContext, sessionManager));
 
   return router;
 };
