+++ app/pages/sections/catalogue-solutions/select/recipient/ui-tests/general.ui.test.js
@@ -4,7 +4,7 @@ import { extractInnerText } from 'buying-catalogue-library';
 import content from '../manifest.json';
 import { orderApiUrl, solutionsApiUrl } from '../../../../../../config';
 
-const pageUrl = 'http://localhost:1234/order/organisation/order-id/catalogue-solutions/solution/price/recipient';
+const pageUrl = 'http://localhost:1234/order/organisation/order-id/catalogue-solutions/select/solution/price/recipient';
 
 const setCookies = ClientFunction(() => {
   const cookieValue = JSON.stringify({
@@ -105,7 +105,7 @@ test('should render Catalogue-solutions select-recipient page', async (t) => {
     .expect(page.exists).ok();
 });
 
-test('should navigate to /organisation/order-id/catalogue-solutions/solution/price when click on backlink', async (t) => {
+test('should navigate to /organisation/order-id/catalogue-solutions/select/solution/price when click on backlink', async (t) => {
   await pageSetup();
   await t.navigateTo(pageUrl);
 
@@ -114,7 +114,7 @@ test('should navigate to /organisation/order-id/catalogue-solutions/solution/pri
   await t
     .expect(goBackLink.exists).ok()
     .click(goBackLink)
-    .expect(getLocation()).eql('http://localhost:1234/order/organisation/order-id/catalogue-solutions/solution/price');
+    .expect(getLocation()).eql('http://localhost:1234/order/organisation/order-id/catalogue-solutions/select/solution/price');
 });
 
 test('should render the title', async (t) => {
