+++ app/pages/sections/catalogue-solutions/select/solution/ui-tests/general.ui.test.js
@@ -4,7 +4,7 @@ import { extractInnerText } from 'buying-catalogue-library';
 import content from '../manifest.json';
 import { solutionsApiUrl, orderApiUrl } from '../../../../../../config';
 
-const pageUrl = 'http://localhost:1234/order/organisation/order-id/catalogue-solutions/solution';
+const pageUrl = 'http://localhost:1234/order/organisation/order-id/catalogue-solutions/select/solution';
 
 const setCookies = ClientFunction(() => {
   const cookieValue = JSON.stringify({
@@ -154,7 +154,7 @@ test('should render the Continue button', async (t) => {
     .expect(await extractInnerText(button)).eql(content.continueButtonText);
 });
 
-test('should redirect to /organisation/order-id/catalogue-solutions/solution/price when a solution is selected', async (t) => {
+test('should redirect to /organisation/order-id/catalogue-solutions/select/solution/price when a solution is selected', async (t) => {
   await pageSetup(true, true);
   await t.navigateTo(pageUrl);
 
@@ -165,7 +165,7 @@ test('should redirect to /organisation/order-id/catalogue-solutions/solution/pri
   await t
     .click(firstSolution)
     .click(button)
-    .expect(getLocation()).eql('http://localhost:1234/order/organisation/order-id/catalogue-solutions/solution/price');
+    .expect(getLocation()).eql('http://localhost:1234/order/organisation/order-id/catalogue-solutions/select/solution/price');
 });
 
 test('should show the error summary when no solution selected causing validation error', async (t) => {
